Did some HEAVY modifications to the let, const, if, print, doubleAssign code and added a push command for arrays along with seperating floats and ints again

+++ src/keywords/function.ts
@@ -26,7 +26,7 @@ export default function functionKey(
 							break;
 						case "number":
 							assignment.push(`${lineTokens[i].value}:`);
-							assignment.push(`f32`);
+							assignment.push(`usize`);
 							i += 2;
 							break;
 					}
